+++ apps/api/src/app/portfolio/portfolio.controller.ts
@@ -375,11 +375,12 @@ export class PortfolioController {
     @Param('dataSource') dataSource: DataSource,
     @Param('symbol') symbol: string
   ): Promise<PortfolioHoldingResponse> {
-    const holding = await this.portfolioService.getHolding(
+    const holding = await this.portfolioService.getHolding({
       dataSource,
       impersonationId,
-      symbol
-    );
+      symbol,
+      userId: this.request.user.id
+    });
 
     if (!holding) {
       throw new HttpException(
@@ -412,19 +413,19 @@ export class PortfolioController {
       filterByAssetClasses,
       filterByDataSource,
       filterByHoldingType,
-      filterBySearchQuery,
       filterBySymbol,
       filterByTags
     });
 
-    const { holdings } = await this.portfolioService.getDetails({
+    const holdings = await this.portfolioService.getHoldings({
       dateRange,
       filters,
       impersonationId,
+      query: filterBySearchQuery,
       userId: this.request.user.id
     });
 
-    return { holdings: Object.values(holdings) };
+    return { holdings };
   }
 
   @Get('investments')
@@ -453,7 +454,8 @@ export class PortfolioController {
       filters,
       groupBy,
       impersonationId,
-      savingsRate: this.request.user?.Settings?.settings.savingsRate
+      savingsRate: this.request.user?.Settings?.settings.savingsRate,
+      userId: this.request.user.id
     });
 
     if (
@@ -622,11 +624,12 @@ export class PortfolioController {
     @Param('dataSource') dataSource: DataSource,
     @Param('symbol') symbol: string
   ): Promise<PortfolioHoldingResponse> {
-    const holding = await this.portfolioService.getHolding(
+    const holding = await this.portfolioService.getHolding({
       dataSource,
       impersonationId,
-      symbol
-    );
+      symbol,
+      userId: this.request.user.id
+    });
 
     if (!holding) {
       throw new HttpException(
@@ -643,7 +646,10 @@ export class PortfolioController {
   public async getReport(
     @Headers(HEADER_KEY_IMPERSONATION.toLowerCase()) impersonationId: string
   ): Promise<PortfolioReportResponse> {
-    const report = await this.portfolioService.getReport(impersonationId);
+    const report = await this.portfolioService.getReport({
+      impersonationId,
+      userId: this.request.user.id
+    });
 
     if (
       this.configurationService.get('ENABLE_FEATURE_SUBSCRIPTION') &&
@@ -672,11 +678,12 @@ export class PortfolioController {
     @Param('dataSource') dataSource: DataSource,
     @Param('symbol') symbol: string
   ): Promise<void> {
-    const holding = await this.portfolioService.getHolding(
+    const holding = await this.portfolioService.getHolding({
       dataSource,
       impersonationId,
-      symbol
-    );
+      symbol,
+      userId: this.request.user.id
+    });
 
     if (!holding) {
       throw new HttpException(
@@ -707,11 +714,12 @@ export class PortfolioController {
     @Param('dataSource') dataSource: DataSource,
     @Param('symbol') symbol: string
   ): Promise<void> {
-    const holding = await this.portfolioService.getHolding(
+    const holding = await this.portfolioService.getHolding({
       dataSource,
       impersonationId,
-      symbol
-    );
+      symbol,
+      userId: this.request.user.id
+    });
 
     if (!holding) {
       throw new HttpException(