+++ libs/common/src/lib/routes/interfaces/internal-route.interface.ts
@@ -1,9 +1,9 @@
 import { User } from '@ghostfolio/common/interfaces';
 
-export interface IRoute {
+export interface InternalRoute {
   excludeFromAssistant?: boolean | ((aUser: User) => boolean);
   path: string;
   routerLink: string[];
-  subRoutes?: Record<string, IRoute>;
+  subRoutes?: Record<string, InternalRoute>;
   title: string;
 }