+++ CHANGELOG.md
@@ -14,7 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Moved the main content of the holding detail dialog to a new overview tab
+- Introduced fuzzy search for the holdings of the assistant
+- Introduced fuzzy search for the quick links of the assistant
 - Improved the search results of the assistant to only display categories with content
+- Enhanced the sitemap to dynamically compose public routes
 - Renamed `Account` to `account` in the `Order` database schema
 - Improved the language localization for German (`de`)